[test][integration] Common jobs webhook package.
