--- test/integration/controller/jobs/jobset/jobset_webhook/suite_test.go
@@ -1,81 +1,0 @@
-/*
-Copyright 2023 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package jobset_webhook
-
-import (
-	"context"
-	"path/filepath"
-	"testing"
-
-	"github.com/onsi/ginkgo/v2"
-	"github.com/onsi/gomega"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-
-	"sigs.k8s.io/kueue/pkg/constants"
-	"sigs.k8s.io/kueue/pkg/controller/jobframework"
-	"sigs.k8s.io/kueue/pkg/controller/jobs/jobset"
-	"sigs.k8s.io/kueue/test/integration/framework"
-)
-
-var (
-	cfg           *rest.Config
-	k8sClient     client.Client
-	ctx           context.Context
-	fwk           *framework.Framework
-	crdPath       = filepath.Join("..", "..", "..", "..", "..", "..", "config", "components", "crd", "bases")
-	jobsetCrdPath = filepath.Join("..", "..", "..", "..", "..", "..", "dep-crds", "jobset-operator")
-	webhookPath   = filepath.Join("..", "..", "..", "..", "..", "..", "config", "components", "webhook")
-)
-
-func TestAPIs(t *testing.T) {
-	gomega.RegisterFailHandler(ginkgo.Fail)
-
-	ginkgo.RunSpecs(t,
-		"JobSet Webhook Suite",
-	)
-}
-
-var _ = ginkgo.BeforeSuite(func() {
-	fwk = &framework.Framework{
-		CRDPath:     crdPath,
-		DepCRDPaths: []string{jobsetCrdPath},
-		WebhookPath: webhookPath,
-	}
-	cfg = fwk.Init()
-	ctx, k8sClient = fwk.SetupClient(cfg)
-})
-
-var _ = ginkgo.AfterSuite(func() {
-	fwk.Teardown()
-})
-
-func managerSetup(opts ...jobframework.Option) framework.ManagerSetup {
-	return func(ctx context.Context, mgr manager.Manager) {
-		reconciler := jobset.NewReconciler(
-			mgr.GetClient(),
-			mgr.GetEventRecorderFor(constants.JobControllerName),
-			opts...)
-		err := jobset.SetupIndexes(ctx, mgr.GetFieldIndexer())
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		err = reconciler.SetupWithManager(mgr)
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		err = jobset.SetupJobSetWebhook(mgr, opts...)
-		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-	}
-}
