+++ test/integration/controller/jobs/jobset/suite_test.go
@@ -46,7 +46,6 @@ var (
 	fwk           *framework.Framework
 	crdPath       = filepath.Join("..", "..", "..", "..", "..", "config", "components", "crd", "bases")
 	jobsetCrdPath = filepath.Join("..", "..", "..", "..", "..", "dep-crds", "jobset-operator")
-	webhookPath   = filepath.Join("..", "..", "..", "..", "..", "config", "components", "webhook")
 )
 
 func TestAPIs(t *testing.T) {
