+++ test/integration/webhook/core/localqueue_test.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package webhook
+package core
 
 import (
 	"github.com/onsi/ginkgo/v2"
